Project4_rest_api/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-comments-xtd==2.10.6
django-contrib-comments==2.2.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
docopt==0.6.2
docutils==0.21.2
idna==3.10
pipreqs==0.4.13
psycopg2-binary==2.9.10
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
yarg==0.1.10
